requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot>=21.6
groq
supabase==2.9.0
python-dotenv
pydantic